+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `laravel-xss-protection` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.2.0 - 2022-02-03
+
+- added `skipKeyWhen` method
+
+## 1.1.0 / 1.1.1 - 2022-02-02
+
+- added `MaliciousInputFound` event
+
+## 1.0.0 - 2022-02-02
 
 - initial release